+++ examples/opengl.rs
@@ -1,6 +1,6 @@
 use libmpv2::{
-    render::{OpenGLInitParams, RenderContext, RenderParam, RenderParamApiType},
     Mpv,
+    render::{OpenGLInitParams, RenderContext, RenderParam, RenderParamApiType},
 };
 use std::{env, ffi::c_void};
 
@@ -44,7 +44,8 @@ fn main() {
         .register_custom_event::<UserEvent>()
         .unwrap();
 
-    mpv.event_context_mut().disable_deprecated_events().unwrap();
+    let mut ev_ctx = mpv.create_event_context();
+    ev_ctx.disable_deprecated_events().unwrap();
 
     let event_sender = event_subsystem.event_sender();
     render_context.set_update_callback(move || {
@@ -54,7 +55,7 @@ fn main() {
     });
 
     let event_sender = event_subsystem.event_sender();
-    mpv.event_context_mut().set_wakeup_callback(move || {
+    ev_ctx.set_wakeup_callback(move || {
         event_sender
             .push_custom_event(UserEvent::MpvEventAvailable)
             .unwrap();
@@ -75,7 +76,7 @@ fn main() {
                         window.gl_swap_window();
                     }
                     UserEvent::MpvEventAvailable => loop {
-                        match mpv.event_context_mut().wait_event(0.0) {
+                        match ev_ctx.wait_event(0.0) {
                             Some(Ok(libmpv2::events::Event::EndFile(_))) => {
                                 break 'render;
                             }